PO-1993 fix tests
